Release 1.0.0.beta.0
This commit represents an effort made by the Checkout team to build a brand new SDK for Ruby. It ships a
new instantiation layer that makes easier a less confusing to set up the SDK. From the point of view of
usability, the new version follows similar principles, although major interfaces are now split/duplicated
to reference different data schemes which makes it fully compatible with `default` (NAS) and `Previous` (ABC)
account structures. This separation is statically forced during instantiation.

The SDK supports the following modules:
* Tokens (default)
* Instruments (default)
* Customers (default)
* Disputes (default)
* Payments (default)
* Alternative Payments (default)
* Forex (default)
* Sessions (default)

`README.md` was updated to match the new setup. Code of conduct and license were reviewed.